# src/swagger.yaml
openapi: 3.0.0
info:
  title: Villageois Automation API
  description: API pour automatiser les actions des villageois dans un jeu.
  version: 1.0.0

servers:
  - url: http://10.110.5.153:3000/api
    description: Serveur de développement local

paths:
  /villageois/move:
    post:
      summary: Déplacer un villageois
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: ID du villageois juste les 2 premier caractères
                  example: "17e9cdb2-6bb1-484e-ad06-5f49c47e2034"
                dest_x:
                  type: number
                  description: Coordonnée X de la destination
                  example: 10
                dest_y:
                  type: number
                  description: Coordonnée Y de la destination
                  example: 15
      responses:
        '200':
          description: Villageois déplacé avec succès
        '404':
          description: Villageois non trouvé

  /villageois/movetoressource:
    post:
      summary: Déplacer un villageois vers la ressource la plus proche
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_villageois:
                  type: string
                  description: ID du villageois juste les 2 premier caractères
                  example: "17e9cdb2-6bb1-484e-ad06-5f49c47e2034"
                x:
                  type: number
                  description: x villageois
                  example: 10
                y:
                  type: number
                  description: y villageois
                  example: 15
                ressourceName:
                  type: string
                  description: ressource name
                  example: "FER"
      responses:
        '200':
          description: Villageois se déplace

  /construction/build:
    post:
      summary: Construire un bâtiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: ID du villageois juste les 2 premier caractères
                  example: "17"
                batiment:
                  type: string
                  description: Type de bâtiment à construire
                  example: "CARRIERE"
      responses:
        '200':
          description: Construction lancée avec succès
        '404':
          description: Villageois non trouvé

  /construction/move-and-build:
    post:
      summary: Déplacer un villageois et construire un bâtiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: ID du villageois juste les 2 premier caractères
                  example: "17"
                batiment:
                  type: string
                  description: Type de bâtiment à construire
                  example: "CARRIERE"
                dest_x:
                  type: number
                  description: Coordonnée X de la destination
                  example: 10
                dest_y:
                  type: number
                  description: Coordonnée Y de la destination
                  example: 15
      responses:
        '200':
          description: Villageois se déplace et commence la construction
        '404':
          description: Villageois non trouvé

  /recolte/recolte:
    post:
      summary: Récolter une ressource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: ID du villageois juste les 2 premier caractères
                  example: "17e9cdb2-6bb1-484e-ad06-5f49c47e2034"
                ressource:
                  type: string
                  description: Ressource à récolter
                  example: "BOIS"
      responses:
        '200':
          description: Récolte lancée avec succès
        '404':
          description: Villageois non trouvé

  /recolte/pause:
    post:
      summary: Mettre en pause un villageois
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: ID du villageois juste les 2 premier caractères
                  example: "17e9cdb2-6bb1-484e-ad06-5f49c47e2034"
      responses:
        '200':
          description: Villageois mis en pause avec succès
        '404':
          description: Villageois non trouvé